+++ src/components/index.ts
@@ -4,7 +4,6 @@ import RestaurantAddModal from "./RestaurantAddModal";
 import CustomSelect from "./CustomSelect";
 import RestaurantCardList from "./RestaurantCardList";
 import RestaurantCard from "./RestaurantCard";
-import RestaurantAddForm from "./RestaurantAddForm";
 
 customElements.define("lunch-header", LunchHeader, { extends: "header" });
 customElements.define("modal-root", ModalRoot);
@@ -14,6 +13,3 @@ customElements.define("restaurant-card-list", RestaurantCardList, {
   extends: "ul",
 });
 customElements.define("restaurant-card", RestaurantCard, { extends: "li" });
-customElements.define("restaurant-add-form", RestaurantAddForm, {
-  extends: "form",
-});
